Updates regarding OpenMP
